+++ src/briefy/leica/tasks/__init__.py
@@ -6,6 +6,7 @@ from briefy.leica.db import db_configure
 from briefy.leica.db import Session
 from briefy.leica.log import tasks_logger as logger
 from briefy.leica.tasks.assignment import move_assignments_awaiting_assets
+from briefy.leica.tasks.assignment import notify_late_submissions
 from briefy.leica.tasks.order import move_orders_accepted
 from briefy.leica.tasks.pool import move_assignments_to_pool
 
@@ -29,6 +30,11 @@ def main():
             logger.info('Start: moving assignments to Awaiting Assets.')
             move_assignments_awaiting_assets()
             logger.info('End: moving assignments to Awaiting Assets.')
+
+        with transaction.manager:
+            logger.info('Start: notifying Assignment late submissions.')
+            notify_late_submissions()
+            logger.info('End: notifying Assignment late submissions..')
 
         with transaction.manager:
             logger.info('Start: moving orders to accepted.')